// src/api/planejador/models/PedidoResponseDTO.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ItemResDto } from './ItemResDto';

export type PedidoResponseDTO = {
  /**
   * @type number
   */
  id: number;
  /**
   * @type string
   */
  codigo: string;
  /**
   * @type string, date-time
   */
  dataEntrega: Date;
  /**
   * @type boolean
   */
  isfake: boolean;
  /**
   * @type number
   */
  lote: number;
  /**
   * @type boolean
   */
  processado: boolean;
  /**
   * @type object
   */
  item: ItemResDto;
};